Move linux listProc() to subtree_linux
